discord/requirements.txt
requests
disnake
fastapi
uvicorn
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
